kafka-spark-docker/consumer/requirements.txt
kafka-python==2.0.2
pyspark==3.5.1